# services/dispatch/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dispatch Service",
    "version": "1.0.0",
    "description": "Trip aggregate and driver dispatch service. Handles ride requests, assigns drivers, and publishes Domain Events to Kafka."
  },
  "paths": {
    "/request-ride": {
      "post": {
        "summary": "Request a ride",
        "operationId": "requestRide",
        "description": "Publishes a TripRequested Domain Event to Kafka and returns the generated trip_id. The Dispatch Service self-consumes the event and assigns a driver, publishing a TripAssigned event.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestRideRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Ride request accepted — TripRequested event published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestRideResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds 64 KB"
          },
          "422": {
            "description": "Validation error — missing or invalid fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v3/api-docs": {
      "get": {
        "summary": "OpenAPI specification",
        "operationId": "getApiDocs",
        "description": "Returns the OpenAPI 3.x JSON specification generated by springdoc-openapi",
        "responses": {
          "200": {
            "description": "OpenAPI 3.x JSON spec",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RequestRideRequest": {
        "type": "object",
        "required": ["riderId", "pickupLocation"],
        "properties": {
          "riderId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "description": "Identifier of the requesting rider"
          },
          "pickupLocation": {
            "$ref": "#/components/schemas/PickupLocation"
          }
        }
      },
      "PickupLocation": {
        "type": "object",
        "required": ["latitude", "longitude"],
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double",
            "minimum": -90.0,
            "maximum": 90.0,
            "description": "WGS-84 latitude in decimal degrees"
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "minimum": -180.0,
            "maximum": 180.0,
            "description": "WGS-84 longitude in decimal degrees"
          }
        }
      },
      "RequestRideResponse": {
        "type": "object",
        "required": ["tripId"],
        "properties": {
          "tripId": {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the newly created Trip aggregate"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "example": "UP"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "message": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
}